refactor to participants.py
