+++ ephios/core/forms/events.py
@@ -18,9 +18,9 @@ from dynamic_preferences.forms import PreferenceForm
 from guardian.shortcuts import assign_perm, get_objects_for_user, get_users_with_perms, remove_perm
 from recurrence.forms import RecurrenceField
 
-from ephios.core import signup
 from ephios.core.dynamic_preferences_registry import event_type_preference_registry
 from ephios.core.models import Event, EventType, LocalParticipation, Shift, UserProfile
+from ephios.core.signup.methods import enabled_signup_methods
 from ephios.core.widgets import MultiUserProfileWidget
 from ephios.extra.crispy import AbortLink
 from ephios.extra.permissions import get_groups_with_perms
@@ -171,7 +171,7 @@ class ShiftForm(forms.ModelForm):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        signup_methods = list(signup.enabled_signup_methods())
+        signup_methods = list(enabled_signup_methods())
         if self.instance and (method_slug := self.instance.signup_method_slug):
             if method_slug not in map(operator.attrgetter("slug"), signup_methods):
                 signup_methods.append(self.instance.signup_method)
