--- ephios/core/signup/__init__.py
@@ -1,1 +1,0 @@
-from .methods import *  # no-qa
