+++ ephios/core/signup/methods.py
@@ -3,7 +3,6 @@ import functools
 import logging
 from argparse import Namespace
 from collections import OrderedDict
-from datetime import date
 from typing import List, Optional
 
 from crispy_forms.bootstrap import FormActions
@@ -11,26 +10,23 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import HTML, Field, Layout
 from django import forms
 from django.contrib import messages
-from django.contrib.auth import get_user_model
 from django.core.exceptions import PermissionDenied, ValidationError
 from django.db import transaction
-from django.db.models import Q, QuerySet
+from django.db.models import Q
 from django.shortcuts import redirect
 from django.template import Context, Template
 from django.template.defaultfilters import yesno
-from django.urls import reverse
 from django.utils import formats, timezone
 from django.utils.functional import SimpleLazyObject, cached_property, classproperty
-from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
 from django.views.generic import FormView
 
-from ephios.core.models import AbstractParticipation, LocalParticipation, Qualification, Shift
+from ephios.core.models import AbstractParticipation, Shift
 from ephios.extra.widgets import CustomSplitDateTimeWidget
 
-from ..models.events import PlaceholderParticipation
 from ..signals import participant_from_request, register_signup_methods
 from .disposition import BaseDispositionParticipationForm
+from .participants import AbstractParticipant
 
 logger = logging.getLogger(__name__)
 
@@ -52,106 +48,6 @@ def signup_method_from_slug(slug, shift=None, event=None):
     raise ValueError(_("Signup Method '{slug}' was not found.").format(slug=slug))
 
 
-@dataclasses.dataclass(frozen=True)
-class AbstractParticipant:
-    first_name: str
-    last_name: str
-    qualifications: QuerySet = dataclasses.field(hash=False)
-    date_of_birth: Optional[date]
-    email: Optional[str]  # if set to None, no notifications are sent
-
-    def get_age(self, today: date = None):
-        if self.date_of_birth is None:
-            return None
-        today, born = today or date.today(), self.date_of_birth
-        return today.year - born.year - ((today.month, today.day) < (born.month, born.day))
-
-    def __str__(self):
-        return f"{self.first_name} {self.last_name}"
-
-    def new_participation(self, shift):
-        raise NotImplementedError
-
-    def participation_for(self, shift):
-        """Return the participation object for a shift. Return None if it does not exist."""
-        raise NotImplementedError
-
-    def all_participations(self):
-        """Return all participations for this participant"""
-        raise NotImplementedError
-
-    @functools.lru_cache
-    def collect_all_qualifications(self) -> set:
-        return Qualification.collect_all_included_qualifications(self.qualifications)
-
-    def has_qualifications(self, qualifications):
-        return set(qualifications) <= self.collect_all_qualifications()
-
-    def reverse_signup_action(self, shift):
-        raise NotImplementedError
-
-    def reverse_event_detail(self, event):
-        raise NotImplementedError
-
-    @property
-    def icon(self):
-        return mark_safe('<span class="fa fa-user"></span>')
-
-
-@dataclasses.dataclass(frozen=True)
-class LocalUserParticipant(AbstractParticipant):
-    user: get_user_model()
-
-    def new_participation(self, shift):
-        return LocalParticipation(shift=shift, user=self.user)
-
-    def participation_for(self, shift):
-        try:
-            return LocalParticipation.objects.get(shift=shift, user=self.user)
-        except LocalParticipation.DoesNotExist:
-            return None
-
-    def all_participations(self):
-        return LocalParticipation.objects.filter(user=self.user)
-
-    def reverse_signup_action(self, shift):
-        return reverse("core:signup_action", kwargs=dict(pk=shift.pk))
-
-    def reverse_event_detail(self, event):
-        return event.get_absolute_url()
-
-
-@dataclasses.dataclass(frozen=True)
-class PlaceholderParticipant(AbstractParticipant):
-    def new_participation(self, shift):
-        return PlaceholderParticipation(
-            shift=shift, first_name=self.first_name, last_name=self.last_name
-        )
-
-    def participation_for(self, shift):
-        try:
-            return PlaceholderParticipation.objects.get(
-                shift=shift, first_name=self.first_name, last_name=self.last_name
-            )
-        except PlaceholderParticipation.DoesNotExist:
-            return None
-
-    def all_participations(self):
-        return AbstractParticipation.objects.none()
-
-    def reverse_signup_action(self, shift):
-        raise NotImplementedError
-
-    def reverse_event_detail(self, event):
-        raise NotImplementedError
-
-    @property
-    def icon(self):
-        return mark_safe(
-            f'<span class="fa fa-user-tag" data-toggle="tooltip" data-placement="left" title="{_("Placeholder")}"></span>'
-        )
-
-
 class ParticipationError(ValidationError):
     pass
 
@@ -166,7 +62,7 @@ def prevent_getting_participant_from_request_user(request):
     request.user = ProtectedUser(lambda: original_user)
 
 
-def get_nonlocal_participant_from_session(request):
+def get_nonlocal_participant_from_request(request):
     for _, participant in participant_from_request.send(sender=None, request=request):
         if participant is not None:
             return participant
@@ -240,15 +136,15 @@ class BaseSignupView(FormView):
     def form_valid(self, form):
         if (choice := self.request.POST.get("signup_choice")) is not None:
             if choice == "sign_up":
-                return self.signup_pressed(form)
+                return self.signup_pressed(**form.cleaned_data)
             if choice == "decline":
-                return self.decline_pressed(form)
+                return self.decline_pressed(**form.cleaned_data)
         return self.form_invalid(form)
 
-    def signup_pressed(self, form):
+    def signup_pressed(self, **signup_kwargs):
         try:
             with transaction.atomic():
-                self.method.perform_signup(self.participant, **form.cleaned_data)
+                self.method.perform_signup(self.participant, **signup_kwargs)
                 messages.success(
                     self.request,
                     self.method.signup_success_message.format(shift=self.shift),
@@ -258,10 +154,10 @@ class BaseSignupView(FormView):
                 messages.error(self.request, self.method.signup_error_message.format(error=error))
         return redirect(self.participant.reverse_event_detail(self.shift.event))
 
-    def decline_pressed(self, form):
+    def decline_pressed(self, **signup_kwargs):
         try:
             with transaction.atomic():
-                self.method.perform_decline(self.participant, **form.cleaned_data)
+                self.method.perform_decline(self.participant, **signup_kwargs)
                 messages.info(
                     self.request, self.method.decline_success_message.format(shift=self.shift)
                 )
