+++ ephios/plugins/basesignup/signup/no_selfservice.py
@@ -2,7 +2,7 @@ from django import forms
 from django.utils.translation import gettext_lazy as _
 
 from ephios.core.models import AbstractParticipation
-from ephios.core.signup import BaseSignupMethod, ParticipationError
+from ephios.core.signup.methods import BaseSignupMethod, ParticipationError
 from ephios.plugins.basesignup.signup.common import render_basic_participation_pills_shift_state
