+++ ephios/plugins/guests/models.py
@@ -7,9 +7,9 @@ from django.urls import reverse
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext_lazy as _
 
-from ephios.core.models import AbstractParticipation, Event
+from ephios.core.models import AbstractParticipation, Event, Qualification
 from ephios.core.models.events import PARTICIPATION_LOG_CONFIG
-from ephios.core.signup import AbstractParticipant, Qualification
+from ephios.core.signup.participants import AbstractParticipant
 from ephios.modellogging.log import register_model_for_logging
